+++ system/txn.h
@@ -1,127 +1,126 @@
-#pragma once 
+#pragma once
 
+#include "database.h"
 #include "global.h"
 #include "helper.h"
-#include "database.h"
 
 class row_t;
 class table_t;
 struct BaseQuery;
 class INDEX;
 
-// each thread has a txn_man. 
+// each thread has a txn_man.
 // a txn_man corresponds to a single transaction.
 
-//For VLL
-enum TxnType {VLL_Blocked, VLL_Free};
+// For VLL
+enum TxnType { VLL_Blocked, VLL_Free };
 
 class Access {
 public:
-	access_t 	type;
-	row_t * 	orig_row;
-	row_t * 	data;
-	row_t * 	orig_data;
-	void cleanup();
+  access_t type;
+  row_t *orig_row;
+  row_t *data;
+  row_t *orig_data;
+  void cleanup();
 #if CC_ALG == TICTOC
-	ts_t 		wts;
-	ts_t 		rts;
+  ts_t wts;
+  ts_t rts;
 #elif CC_ALG == SILO
-	ts_t 		tid;
-	ts_t 		epoch;
+  ts_t tid;
+  ts_t epoch;
 #elif CC_ALG == HEKATON
-	void * 		history_entry;	
+  void *history_entry;
 #endif
-
 };
 
-class txn_man
-{
+class txn_man {
 public:
-	virtual void 	initialize(Database * db, uint32_t thread_id);
-	virtual RC 		run_txn(BaseQuery * m_query) = 0;
-	void 			release();
-	void 			cleanup(RC rc);
-	RC 				finish(RC rc);
-	itemid_t *		index_read(INDEX * index, idx_key_t key, int part_id);
-	void 			index_read(INDEX * index, idx_key_t key, int part_id, itemid_t *& item);
-	row_t * 		get_row(row_t * row, access_t type);
+  virtual void initialize(Database *db, uint64_t thread_id);
+  virtual RC run_txn(BaseQuery *m_query) = 0;
+  void release();
+  void cleanup(RC rc);
+  RC finish(RC rc);
+  itemid_t *index_read(INDEX *index, idx_key_t key, int part_id);
+  void index_read(INDEX *index, idx_key_t key, int part_id, itemid_t *&item);
+  row_t *get_row(row_t *row, access_t type);
 
-	uint64_t 		get_thd_id();
-	Database * 		get_db();
+  uint64_t get_thd_id();
+  Database *get_db();
 
-	void 			set_txn_id(txnid_t txn_id);
-	txnid_t 		get_txn_id();
+  void set_txn_id(txnid_t txn_id);
+  txnid_t get_txn_id();
 
-	void 			set_ts(ts_t timestamp);
-	ts_t 			get_ts();
+  void set_ts(ts_t timestamp);
+  ts_t get_ts();
 
-	void reset(txnid_t txnid) {
-		this->txn_id = txnid;
-		this->abort_cnt = 0;
-	}
+  void reset(txnid_t txnid) {
+    this->txn_id = txnid;
+    this->abort_cnt = 0;
+  }
 
-	Database * 			database;
-	myrand * 			mrand;
-	uint64_t 			abort_cnt;
-	uint32_t 			thread_id;
-	pthread_mutex_t 	txn_lock;
-	row_t * volatile 	cur_row;
+  Database *database;
+  myrand *mrand;
+  uint64_t abort_cnt;
+  uint64_t thread_id;
+  pthread_mutex_t txn_lock;
+  row_t *volatile cur_row;
 
 #if CC_ALG == HEKATON
-	void * volatile history_entry;
+  void *volatile history_entry;
 #endif
-	// [DL_DETECT, NO_WAIT, WAIT_DIE]
-	bool volatile 	lock_ready;
-	bool volatile 	lock_abort; // forces another waiting txn to abort.
-	// [TIMESTAMP, MVCC]
-	bool volatile 	ts_ready; 
-	// [HSTORE]
-	int volatile 	ready_part;
+  // [DL_DETECT, NO_WAIT, WAIT_DIE]
+  bool volatile lock_ready;
+  bool volatile lock_abort; // forces another waiting txn to abort.
+  // [TIMESTAMP, MVCC]
+  bool volatile ts_ready;
+  // [HSTORE]
+  int volatile ready_part;
 
 #if CC_ALG == TICTOC
-	ts_t 			get_max_wts() 	{ return _max_wts; }
-	void 			update_max_wts(ts_t max_wts);
-	ts_t 			last_wts;
-	ts_t 			last_rts;
+  ts_t get_max_wts() { return _max_wts; }
+  void update_max_wts(ts_t max_wts);
+  ts_t last_wts;
+  ts_t last_rts;
 #elif CC_ALG == SILO
-	ts_t 			last_tid;
+  ts_t last_tid;
 #endif
 
-	// For OCC
-	uint64_t 		start_ts;
-	uint64_t 		end_ts;
-	// following are public for OCC
-	int 			row_cnt;
-	int	 			wr_cnt;
-	Access **		accesses;
-	int 			num_accesses_alloc;
+  // For OCC
+  uint64_t start_ts;
+  uint64_t end_ts;
+  // following are public for OCC
+  int row_cnt;
+  int wr_cnt;
+  Access **accesses;
+  int num_accesses_alloc;
 
-	// For VLL
-	TxnType 		vll_txn_type;
+  // For VLL
+  TxnType vll_txn_type;
 
-protected:	
-	void 			insert_row(row_t * row, table_t * table);
+protected:
+  void insert_row(row_t *row, table_t *table);
+
 private:
-	// insert rows
-	uint64_t 		insert_cnt;
-	row_t * 		insert_rows[MAX_ROW_PER_TXN];
-	txnid_t 		txn_id;
-	ts_t 			timestamp;
+  // insert rows
+  uint64_t insert_cnt;
+  row_t *insert_rows[MAX_ROW_PER_TXN];
+  txnid_t txn_id;
+  ts_t timestamp;
 
-	bool _write_copy_ptr;
+  bool _write_copy_ptr;
 #if CC_ALG == TICTOC || CC_ALG == SILO
-	bool 			_pre_abort;
-	bool 			_validation_no_wait;
+  bool _pre_abort;
+  bool _validation_no_wait;
 #endif
 #if CC_ALG == TICTOC
-	bool			_atomic_timestamp;
-	ts_t 			_max_wts;
-	// the following methods are defined in concurrency_control/tictoc.cpp
-	RC				validate_tictoc();
+  bool _atomic_timestamp;
+  ts_t _max_wts;
+  // the following methods are defined in concurrency_control/tictoc.cpp
+  RC validate_tictoc();
 #elif CC_ALG == SILO
-	ts_t 			_cur_tid;
-	RC				validate_silo();
+  ts_t _cur_tid;
+  RC validate_silo();
 #elif CC_ALG == HEKATON
-	RC 				validate_hekaton(RC rc);
+  RC validate_hekaton(RC rc);
 #endif
 };
