+++ workload/distributions.h
@@ -13,51 +13,51 @@
  * Each thread has its own buffer data, so individually the threads generate
  * random distribution.
  *
- * Note that this is also thread-safe. It basically produces n sequences of numbers
- * in which each sequence follows a random distribution.
+ * Note that this is also thread-safe. It basically produces n sequences of
+ * numbers in which each sequence follows a random distribution.
  */
 class RandomNumberGenerator {
 public:
-  explicit RandomNumberGenerator(uint32_t num_dists) {
+  explicit RandomNumberGenerator(uint64_t num_dists) {
     _num_dists = num_dists;
     _buffers = new drand48_data *[num_dists];
     _latches = new pthread_mutex_t[num_dists];
-    for (uint32_t i = 0; i < num_dists; i++) {
+    for (uint64_t i = 0; i < num_dists; i++) {
       _buffers[i] = reinterpret_cast<drand48_data *>(
           _mm_malloc(sizeof(drand48_data), 64));
       pthread_mutex_init(&_latches[i], NULL);
     }
   }
   ~RandomNumberGenerator() {
-    for (uint32_t i = 0; i < _num_dists; i++) {
+    for (uint64_t i = 0; i < _num_dists; i++) {
       pthread_mutex_destroy(&_latches[i]);
       free(_buffers[i]);
     }
     delete[] _buffers;
     delete[] _latches;
   }
-  uint64_t nextInt64(uint32_t dist_id) {
+  uint64_t nextInt64(uint64_t dist_id) {
     int64_t rint64;
     pthread_mutex_lock(&_latches[dist_id]);
     lrand48_r(_buffers[dist_id], &rint64);
     pthread_mutex_unlock(&_latches[dist_id]);
     return static_cast<uint64_t>(rint64);
   }
-  double nextDouble(uint32_t dist_id) {
+  double nextDouble(uint64_t dist_id) {
     double rdouble;
     pthread_mutex_lock(&_latches[dist_id]);
     drand48_r(_buffers[dist_id], &rdouble);
     pthread_mutex_unlock(&_latches[dist_id]);
     return rdouble;
   }
-  void seed(uint32_t dist_id, long value) {
+  void seed(uint64_t dist_id, uint64_t value) {
     pthread_mutex_lock(&_latches[dist_id]);
-    srand48_r(value, _buffers[dist_id]);
+    srand48_r((long)value, _buffers[dist_id]);
     pthread_mutex_unlock(&_latches[dist_id]);
   }
 
 protected:
-  uint32_t _num_dists;
+  uint64_t _num_dists;
   drand48_data **_buffers;
   pthread_mutex_t *_latches;
 };
@@ -80,14 +80,14 @@ protected:
  */
 class ZipfianNumberGenerator {
 public:
-  ZipfianNumberGenerator(uint64_t n, uint32_t num_distributions,
+  ZipfianNumberGenerator(uint64_t n, uint64_t num_distributions,
                          double zipfian_theta)
       : _generator(num_distributions) {
     _the_n = n - 1;
     _theta = zipfian_theta;
     initialize();
   }
-  uint64_t nextZipfInt64(uint32_t dist_id) {
+  uint64_t nextZipfInt64(uint64_t dist_id) {
     double u = _generator.nextDouble(dist_id);
     double uz = u * _zeta_n_theta;
     if (uz < 1) {
@@ -98,13 +98,13 @@ public:
       return 1 + (uint64_t)(_the_n * pow(_eta * u - _eta + 1, _alpha));
     }
   }
-  uint64_t nextRandInt64(uint32_t dist_id) {
+  uint64_t nextRandInt64(uint64_t dist_id) {
     return _generator.nextInt64(dist_id);
   }
-  double nextRandDouble(uint32_t dist_id) {
+  double nextRandDouble(uint64_t dist_id) {
     return _generator.nextDouble(dist_id);
   }
-  void seed(uint32_t dist_id, long value) { _generator.seed(dist_id, value); }
+  void seed(uint64_t dist_id, uint64_t value) { _generator.seed(dist_id, value); }
   ~ZipfianNumberGenerator() = default;
 
 protected:
@@ -115,7 +115,6 @@ protected:
     _eta = (1 - pow(2.0 / _the_n, 1 - _theta)) /
            (1 - _zeta_2_theta / _zeta_n_theta);
   }
-
   static double zeta(uint64_t n, double theta) {
     double sum = 0;
     for (uint64_t i = 1; i <= n; i++) {
@@ -123,7 +122,6 @@ protected:
     }
     return sum;
   }
-
   double _theta;
   uint64_t _the_n;
   double _zeta_n_theta;
